fix: two defects the host tests could not see
Both surfaced on the first firmware build after Tasks 2 and 3, and neither
is reachable from tests/host:

- page_name() in ui_app.cpp has no default case, so adding PageId::NowPlaying
  in Task 2 broke the build under -Werror=switch. Host tests never compile
  ui_app.cpp.
- format_track_time() printed uint32_t with %u. uint32_t is `unsigned int`
  on the host and `unsigned long` on xtensa, so the host build is clean and
  the firmware build fails under -Werror=format. Casting at the call site
  keeps one format string correct on both.

The lesson is the schedule, not the two lines: the plan deferred the first
firmware build to Task 5, and these sat undetected across two green test
runs. Every remaining task builds the firmware.

Also drops a protocol-specific dB value from a ui_data.hpp comment, matching
the same call made in media_registry.hpp earlier.

+++ components/ui/include/ui_data.hpp
@@ -1550,21 +1550,31 @@ constexpr int volume_overlay_fill_width(float volume) {
 // into hours - an hour-long track shows 61:01, which is unambiguous and needs
 // no third field. Truncates: 1:59.9 is still 1:59, because a clock that
 // reaches 2:00 before the track does reads as broken.
+//
+// The casts are load-bearing, not decoration: uint32_t is `unsigned int` on
+// the host and `unsigned long` on xtensa, so a bare %u compiles clean under
+// the host tests and fails the firmware build outright under -Werror=format.
+// Casting at the call site keeps one format string correct on both, which
+// PRIu32 would also do at the cost of breaking the string into fragments.
+// This cost a firmware build to find; every printf-family call reached from a
+// header compiled for both targets has the same trap in it.
 inline std::string format_track_time(uint32_t milliseconds) {
   const uint32_t total_seconds = milliseconds / 1000;
   char buffer[16];
-  std::snprintf(buffer, sizeof(buffer), "%u:%02u", total_seconds / 60,
-                total_seconds % 60);
+  std::snprintf(buffer, sizeof(buffer), "%u:%02u",
+                static_cast<unsigned>(total_seconds / 60),
+                static_cast<unsigned>(total_seconds % 60));
   return buffer;
 }
 
 // The digits only - the '%' is drawn separately, because the 128px face is
 // rlcd_digits_128.c and has no glyph for it.
 //
-// Three distinct returns, deliberately: "MUTE" for a real muted source
-// (AirPlay's -144 dB), "0" for a source turned all the way down but not
-// muted, and an empty string for a source that has not reported a level at
-// all. The overlay does not open on the empty case.
+// Three distinct returns, deliberately: "MUTE" for a source that says it is
+// muted, "0" for one turned all the way down but not muted, and an empty
+// string for one that has not reported a level at all. The overlay does not
+// open on the empty case. How a protocol spells "muted" on the wire is the
+// publishing module's business - see NowPlaying::muted.
 inline std::string volume_percent_text(float volume, bool muted) {
   if (volume < 0.0f) return {};
   if (muted) return "MUTE";
